[MLIR] Handle exceptions (#447)
**Context:** Exceptions raised during run time would trigger
segmantation faults.

**Description of the Change:** Adds MLIR pass to check for errors
returned from asynchronous functions

**Benefits:** No segfaults.

**Possible Drawbacks:** One thing I noticed is that even without the
exception handling transformation, it looks like when asynchronous
execution happens in parallel pytests, there is a non-zero probability
of failure. This gets reduced the less number of parallel tests are run.
I will ask the community and see if there is a way to further diagnose
this issue.

**Related GitHub Issues:** This PR does not implement the transformation
that would propagate the exception handling algorithm up the call graph.
This is because inlining guarantees that all calls to quantum operations
happen in the "main" method. But this is something that we might
consider in the future, or do the runtime approach instead. Things that
are left as TODOs:

* Use a stack and a mutex to preserve error messages from multiple
threads
* Before this PR, the runtime did not raised errors when
`qml.CNOT(wires=[x, y])` given `x == y`. There might be other places
where we need to raise these exceptions.
 * De-allocate device and quantum register when an exception happens.
 
[sc-52638]

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>