[Frontend] Execute tapes in cuda-quantum (#477)
**Context:** A map from quantum tapes to CUDA-quantum's Python
interface.

**Description of the Change:** 
* Adds a CUDA-quantum device
* Adds a map from Catalyst JAXPR to CUDA-quantum's Python interface.

In this PR we add a CUDA-quantum device that describes the legal
operations for CUDA-quantum kernels via the Python interface and we add
a way to execute tapes in this CUDA-quantum device. The CUDA-quantum
device targets `qpp-cpu` by default. Execution on CUDA-quantum's API is
driven by a custom JAX interpreter. A JAX interpreter will iterate over
Catalyst JAXPR but instead of executing Catalyst's instructions, it will
execute equivalent instructions in CUDA-quantum's API.

**Benefits:** Users can execute tapes via CUDA-quantum's Python
interface.

**Possible Drawbacks:** Due to symbol conflicts (possibly in OpenMP)
running `import cudaq` before running any kokkos kernel will result in a
segfault.

CUDA-quantum's operations that are not reachable from tapes:
* `ch`
* `sdg`
* `tdg`
* `cs`
* `ct`
* `r1`

Catalyst's operations that are unimplemented:
```
    zne_p: unimplemented_impl,
    qunitary_p: unimplemented_impl,
    hermitian_p: unimplemented_impl,
    tensorobs_p: unimplemented_impl,
    var_p: unimplemented_impl,
    probs_p: unimplemented_impl,
    cond_p: unimplemented_impl,
    while_p: unimplemented_impl,
    for_p: unimplemented_impl,
    grad_p: unimplemented_impl,
    func_p: unimplemented_impl,
    jvp_p: unimplemented_impl,
    vjp_p: unimplemented_impl,
    print_p: unimplemented_impl,
```

[sc-55098]

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Josh Izaac <josh146@gmail.com>