Rework QJIT class into distinct compilation stages (#531)
This is part 2 of a refactor started in #529. 

The QJIT class is reworked into 5 distinct compilation stages:
- pre-compilation (like autograph)
- capture (jaxpr generation)
- ir-generation (mlir generation)
- compilation (llvm and binary code generation - cannot be split up
since this happens in the compiler driver)
- execution

The class is also streamlined by using a new compilation cache to handle
previously compiled functions and signature lookups.

One point of contention might be the results produced by the split of
the `trace_to_mlir` function, which have been simplified and need to be
double checked against #520. EDIT:
https://github.com/PennyLaneAI/catalyst/pull/531/commits/c71c322c1d76a0373d41e847d5883ab68b937298
should address this concern

[sc-57014]

closes #268 
closes #520