Fix wheel failures
One issue that is fixed is that the lightning pip package and the lightning
C++ library shipped with the runtime are not always in synce. This commit
attempts to remedy that by allowing a prebuilt development wheel for
lightning to be installed automatically when installing Catalyst via make.
Using pip install directly is not recommended because an extra flag is needed
to allow fetching wheels from TestPyPI, and this cannot be supplied from
within the package requirements (setup.py) itself.

In order to depend on development versions of PennyLane or Lightning during
Catalyst development, the .dep-versions file simply needs to be updated:
- for PL with the commit hash on master
- for LQ with the version tag (ending in .dev) corresponding to a specific PR
Additionally for LQ the wheels and runtime MakeFile need to be updated with
the commit hash on corresponding to said PR, but this is a temporary measure.

Temporarily skip one test that is failing due to lack of GlobalPhase support.

Update plugin test matrix action for stable to use latest_release tag in lightning,
since it is okay for stable entries to fail with development versions.