Fix `distutils` import across environments (#667)
This PR resolves issues around the availability of the `sysconfig`
module, and by extension the `distutils` module. `distutils` is present
both in the standard library as well as in `setuptools`. The standard
library module has been removed in Python 3.12, although the `sysconfig`
submodule has been preserved. Additionally, a `setuptools` hijacks the
import process of `distutils` and replaces it with its own
implementation. The provided fix is the following:
- Explicitly import `distutils` from `setuptools` to avoid any
ambiguities.
- Use `pip wheel` instead of `python setup.py` to guarantee the build
environment is set up according to the `pyproject.toml` specification
before building the wheel, such as ensuring `setuptools` is available.

Note that using `sysconfig` from the standard library did not appear to
be compatible for the [context we use it
in](https://github.com/PennyLaneAI/catalyst/blob/0beb6ae552289ef4ad8b5a9c3d56651f983619bf/setup.py#L180).


[sc-61185]