Remove deprecated `qfunc` decorator (#679)
As part of the Frontend refactor this PR removes the deprecated `@qfunc`
decorator that has remained in use for internal testing (specifically
lit testing). However, its use is actually very misleading since it
bypasses a lot of the new infrastructure around devices and
decomposition. Tests have been updated or removed.

[sc-59509]