On-Board check-catalyst workflow to use GitHub Large Runners (#675)
**Context:** 
When high volume of jobs are being queued on CI. It is possible that it
will take a while before a job is picked up. To get around this, we have
provisioned an organization runner group of GitHub Large Runners. Since
these are paid runners, we retain the ability to scale this up to a
value suitable for us during periods of high usage.


**Description of the Change:**
A new label `urgent` has been introduced to this repo. Whenever a pull
request has this label attached, all the jobs within check-catalyst.yml
will swap over to use the new Large runner group described above.

This PR also bumps the version of actions/checkout and actions/cache
being used. This was done to deal with the deprecation of node16 within
github actions.

**Benefits:**
Lowered queue times for urgent CI workloads to get picked up.

**Possible Drawbacks:**
We can still see jobs being queued for a while if enough PRs have the
`urgent` label at once. To get around this, only add it to PRs that are
truly urgent. Also, we can increase the scaling factor of the group if
needed.

**Related GitHub Issues:**
None.
[sc-61791](https://app.shortcut.com/xanaduai/story/61791/update-catalyst-ci-to-use-large-runners)

[sc-61184]: fix for cache issues