Fix CI issue with Linux wheel builds failing / skipping (#685)
**Context:** A bug was introduced with #675 which caused Linux builds to
be skipped. This PR fixes the issue.

On the GitHub UI, it showed all checks as passing which caused the issue
to slip by.

When `fromJSON` attempts to parse a string that is invalid JSON, it
returns an empty string. This caused the strategy matrix to fail.

**Description of the Change:**
Fixed formatting of format function feeding into fromJSON.

**Benefits:**
Linux jobs won't fail due to invalid workflow definition.

**Possible Drawbacks:**
None.

**Related GitHub Issues:**
None.
[sc-61791](https://app.shortcut.com/xanaduai/story/61791/update-catalyst-ci-to-use-large-runners)