[CI/CD] Use 3.9 to build LLVM/MLIR bindings in build wheel action (#684)
**Description of the Change:** To reduce cache utilization and reduce
quantities of LLVM builds in cache, we use python 3.9 to build LLVM and
MLIR's python bindings for linux x86_64 and macOS x86_64. macOS aarch64
still build LLVM and MLIR bindings with 3.9, 3.10, 3.11, and 3.12
because it involves a bit more work with the setup action. Can be made
as a follow up PR.

**Benefits:** 6 versions of LLVM are made redundant.