Add Github action workflow to notify on main jobs failures (#690)
**Context:**
It is useful to receive notifications whenever a github action workflow
running on mail fails. Github has a slack integration that can be
configured, but that sends notification for every job in any
state(started, cancelled, failed, succeeded). There is an [open
issue](https://github.com/integrations/slack/issues/1563) that hasn't
been solved for 2 years.
Because of this, a manual job that triggers based on failures of
selected job on main is a better approach, and it won't need to add the
notification on each single job.

**Description of the Change:**
A new workflow is added that will send notification to a slack
channel(specified in the webhook configuration in the slack settings).

**Benefits:**
Get notification when specified jobs fails on main

**Possible Drawbacks:**
None

**Related GitHub Issues:**
None