[Frontend] Implement quantum device capabilities as a data structure (#609)
In this PR we implement a dedicated data structure representing quantum
device capabilities. Previously we used toml document IR for this
purpose. We see the following benefits in the new approach:

* Data structure simplifies capability calculations we do for our
QJITDevice. Now the capabilities of QJITDevice and of the third-party
devices we support are represented in the same format.
* We reduce the internal usage of `C(Gate)/Adjoint(Gate)` syntax to
minimum. We still use it only to communicate with PennyLane's API


In subsequent PRs we will use it for program verification.

[sc-59478]

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>