Fix pennylane not updating in latest action
Pip will only install a package if the provided requirement does not
already match an installed package. PennyLane maintains the same version
string throughout one development cycle, which led the package to not be
upgraded from the version specified in the Catalyst requirements.

The same fix is not required for lightning, since lightning changes
its version number on every commit to main.