Adds `catalyst.debugCallback` attribute
`PythonCallOp`s are lowered into LLVM functions. This commit
introduces the change that `PythonCallOp`s are lowered into LLVM
functions which are annotated with the `catalyst.debugCallback`
attribute.

The reason for this is that without this attribute, the only
way to determine if an LLVM call corresponds to a function call to
`@pyregistry` with a third parameter set to zero.

To avoid having to find the definition of the third parameter
(which may have been defined in a different function), we annotate
during the lowering.

This case is likely not going to happen in the current state of
the compiler, since it would depend on outlining. But, it is still
a good change to make since it reduces compilation time having to
traverse the use-def chain.