[Frontend] Support for IsingZZ in Catalyst (#730)
**Context:** We would like to add support for the PennyLane IsingZZ
operation to our compilation architecture. The PennyLane-Lightning
simulator supports this gate, and we would like to be able to support it
without relying on the operation decomposition.

**Description of the Change:** Added IsingZZ to the list of supported
gates. Added tests to confirm the functionality of IsingZZ gate and that
the decomposition is not happening anymore.

**Benefits:** No need to decompose IsingZZ gate into a CNOT and RZ gate

**Possible Drawbacks:** May need more complicated tests to capture the
full functionality of IsingZZ gate

**Related GitHub Issues:** #662

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>