Decompose nested tapes in HybridOps Pt 2 (#728)
**Context:**
An initial path-of-least-resistance implementation of `decompose` was
added in #638, because it was necessary to complete the task at hand in
that PR. However, as it wasn't the main focus of that PR, and it was
merged with the plan to fix several things if possible. This PR
implements the "cleaner", more updated version of the transform. In the
initial implementation:

1. There was no testing for the added decomposition of HybridOp tapes
(high priority)
2. A `visited` property was temporarily added to the HybridOp class to
make things easier to implement, and should be removed (if possible)
3. A lot of the code involved was almost, but not quite, identical to
existing code in PL, and the duplication should be reduced (if possible)

**Description of the Change:**
Tests are added, and the code adjusted to fix any issues that came up.
The code was refactored to use most of the comparable functionality in
PennyLane instead of duplicating it in catalyst, and to remove the need
for the `visited` property on `HybridOps`.

**Benefits:**
We made the changes to make this work more cleanly as discussed during
the review process for #638. We avoid duplicate code. We have tests that
confirm that the decomposition of nested tapes is working as expected.