Fixing `jax.scipy.linalg.expm` complaints regarding undefined symbols. (#752)
**Context:** The original issue #628 raises a bug about
`jax.scipy.linalg.expm` not working with `qml.qjit`. Upon investigation,
two bugs were the cause of this: a batching rule query bug which was
fixed in #733 (and #740), and a undefined symbol bug, raised in a new
issue #747 which this PR addresses.

Bug: `jax.scipy.linalg.expm` does not work 
```python
@qml.qjit
def func(x):
	res = jax.scipy.linalg.expm(x)
	return res

y = jnp.array( [[1.0, 0.0], [0.0, 1.0]] )
x = func(y)


Traceback (most recent call last):
  File "/home/paul.wang/expmfix.py", line 43, in <module>
    x = func(y)
  File "/home/paul.wang/catalyst/frontend/catalyst/jit.py", line 110, in __call__
    requires_promotion = self.jit_compile(args)
  File "/home/paul.wang/catalyst/frontend/catalyst/jit.py", line 171, in jit_compile
    self.compiled_function, self.qir = self.compile()
  File "/home/paul.wang/catalyst/frontend/catalyst/debug/instruments.py", line 143, in wrapper
    return fn(*args, **kwargs)
  File "/home/paul.wang/catalyst/frontend/catalyst/jit.py", line 278, in compile
    compiled_fn = CompiledFunction(shared_object, func_name, restype, self.compile_options)
  File "/home/paul.wang/catalyst/frontend/catalyst/compiled_functions.py", line 132, in __init__
    self.shared_object = SharedObjectManager(shared_object_file, func_name)
  File "/home/paul.wang/catalyst/frontend/catalyst/compiled_functions.py", line 61, in __init__
    self.open()
  File "/home/paul.wang/catalyst/frontend/catalyst/compiled_functions.py", line 65, in open
    self.shared_object = ctypes.CDLL(self.shared_object_file)
  File "/usr/lib/python3.10/ctypes/__init__.py", line 374, in __init__
    self._handle = _dlopen(self._name, mode)
OSError: /tmp/func4eeltbvg/func.so: undefined symbol: blas_dtrsm
```

**Description of the Change:**
1. Added new wrapper definitions for blas core routines required by
`jax.scipy.linalg.expm` in
`frontend/catalyst/utils/libcustom_calls.cpp`. See the detailed bug fix
report below.
2. Added a new file `frontend/test/pytest/test_jax_numerical.py` to hold
tests regarding correctness of qjit handling calls to jaxlib. In the
first expm fix PR #733 (and #740 ) the new test for expm was added to
`test_jax_dynamic_api.py`. This is the inappropriate place, and the test
is moved to the new `test_jax_numerial.py`. Originally the test was
marked as xfail since batching rule query was fixed but undefined symbol
was not. Now that expm is fully supported, the test is no longer marked
as xfail.

**Benefits:** `jax.scipy.linalg.expm` now works

**Possible Drawbacks:** other calls to jaxlib, such as `sqrtm`, still
does not work, and the openblas routines need to be added by hand one at
a time. This might not be the best way to link jaxlib. This is not a
drawback in the sense that this PR makes anything worse, but a note that
something could be improved.

**Related GitHub Issues:** closes #628, closes #747 

[sc-63402]