[Runtime] Add support for Lightning source directory to be used in runtime build (#732)
**Context:** This PR allows Lightning to be cloned and used in-place
with Catalyst's runtime, allowing faster turnaround and development
against both packages. <del>In addition, it overcomes some issues with
Python package issues observed at build time by updating the runtime
pybind11 and requiring Python to be found before proceeding.<\del>

**Description of the Change:** Added `LIGHTNING_SRC_PATH` CMake argument
to specify a full path to a local Lightning repository, and all source
will be used in place for a runtime build.

**Benefits:** Enables faster turnaround for development between
Lightning and Catalyst. Can allow a multi-version clone of each repo and
compilation steps to be validated and tested.

**Possible Drawbacks:**

**Related GitHub Issues:**