Refactor vmap following the decorator callable class pattern (#758)
**Context:** We intend that all Catalyst's functional decorators are
presented uniformly to the user. In this case, the `vmap` decorator was
refactored.

**Description of the Change:** Create a `VmapCallable` class and
instantiate an object of that class from the `vmap` API. Previous code
was moved from the API to the callable class.

**Benefits:** The use of def decorator(): and class DecoratorCallable:
is a common pattern used by Catalyst to implement decorators without
having to excessively define functions in a nested fashion. A consistent
implementation pattern is preferred across all decorators.

**TODO:**

- [x] Add test cases
- [x] Update changelog

[sc-60279]