When generating mlir GradOp, correctly querying const array types from parent numpy arrays (#778)
**Context:** #774 
`jax.grad` fails on a qnode when the qnode has a `QubitUnitary` that
takes in a complex matrix. This is because currently in
`frontend/catalyst/jax_primitives.py`, in `_grad_lowering`, when
producing the `ConstantOp`s from frontend's numpy arrays, the dtype of
the numpy array is not passed in, and the type of the produced
`ConstantOp` and the actual data type is not guaranteed to be the same.

**Description of the Change:**
We now pass in the type of the parent numpy array. Jax provides a
utility function that converts numpy types to mlir types (including the
shape):
https://github.com/google/jax/blob/d6a84cc5f327c789da073a11b1c0d3fbc2a56bdd/jaxlib/hlo_helpers.py#L48

~~A similar usage pattern exists when generating vjp/jvp and is updated
as well.~~ More work is needed for `jvp`/`vjp` to fully support complex
values. We thus do not implement it in this PR.

**Benefits:** Bugfix

**Related GitHub Issues:** closes #774 

[sc-64529]