Standardize the call sigantures for measurement primitives (#790)
**Context:**

Each type of measurement primitive has a slightly different call
signature. This lack of standardization means that each type of
primitive needs custom handling. That custom handling makes conversion
and binding of measurements fragile, error-prone, and difficult to
extend.

**Description of the Change:**

Standardizes the call signatures of each measurement primitive to
contain `obs` as the first positional keyword arguments, and `shots` and
`shape` as keyword arguments.

While the call signatures differ the order of `shots` and `shape`, both
of these values should always be passed as keyword arguments due to the
fact they are not traceable data. Differing orders should not effect any
use cases.

**Benefits:**

Able to substitute and dispatch between all the different types of
measurement primitives.

**Possible Drawbacks:**

The order of keyword arguments are still not standardized.

**Related GitHub Issues:**

[sc-65000]