Catalyst and PL Control unification (#771)
**Context:**
Restructure `catalyst.ctrl` to inherit from the PL Controlled class.

This provides feature parity for single operations. For quantum
functions, however, this support is not fully compatible with PL because
PL only instantiates the `Controlled` and `ControlledOp` classes of a
single base operation and general operator and returns a list of
`Controlled` instances while Catalyst intends to return a single
controlled instance with a list of ops nested inside of it.


**Description of the Change:**
- Rename the current `QCtrl(HybirdOp)` class to `HybridCtrl(HybridOp)`.
This class provides support for computing the control of single-op and
quantum functions (multiple-ops).
- Create new classes `catalyst.ControlledOp`, `catalyst.ControlledBase`
derived from `catalyst.Controlled` to provide feature parity with PL for
single-operation and single-operator. The tracer was also updated to
make this class QJIT-compatible.
- Create a new wrapper class `CtrlCallable` to produce a QJIT-compatible
ctrl instance.
- Update `catalyst.ctrl` to create a `CtrlCallable` instance based on
the input base operation/callable.
- Add PL's Controlled unit tests to check the feature parity of
`catalyst.ctrl`.

**Benefits:**
- Better feature parity of `catalyst.ctrl` with PL's ctrl for single
ops.
- Allow `catalyst.ctrl` to be used outside the QJIT or QNode context
similar to the PL ctrl method.

**Possible Drawbacks:**

**Related Tickets:**
[sc-60924]
[sc-60925]

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>