Added a new GradientOpInterface for common attributes in GradOp, JVPOp and VJPOp in the mlir cpp API (#800)
**Context:** VJP, JVP and Grad should all have a method attribute. An
interface for it would be nice. Otherwise we would need a complicated
"visitor"-style check for all three types like that in #789.

**Description of the Change:**
- Added a new interface `catalyst::gradient::GradientOpInterface` in the
gradient dialect, along with necessary registry/cmake boilerplates
- `catalyst::gradient::GradOp`, `JVPOp`, and `VJPOp` now inherits traits
in this new `GradientOpInterface` (right now there is only a
`getMethod()` method, returning "auto"/"fd")
- The complicated visitor check for grad/jvp/vjp in #789 is replaced
with utilization of this new interface

**Benefits:** There are operations that could potentially be used as
`GradOp`, `JVPOp` or `VJPOp`. When trying to get the gradient method,
instead of doing
```C++
        auto gradOp = dyn_cast<GradOp>(op);
        auto jvpOp = dyn_cast<JVPOp>(op);
        auto vjpOp = dyn_cast<VJPOp>(op);

        llvm::StringRef MethodName;
        if (gradOp)
            MethodName = gradOp.getMethod();
        else if (jvpOp)
            MethodName = jvpOp.getMethod();
        else if (vjpOp)
            MethodName = vjpOp.getMethod();
```
to identify which op it actually is and protect against segfaults
(calling `nullptr.getMethod()`), in the new interface we just do
```C++
        auto gradOpInterface = cast<GradientOpInterface>(op);
        llvm::StringRef MethodName = gradOpInterface.getMethod();
```

**Related GitHub Issues:** closes #794 

[sc-65342]