[Frontend] Update execution config (#770)
**Context:**
The expectation wrt the PennyLane API is that there are two
`TransformPrograms` applied to the tapes before execution begins. These
are from the `qnode` and the `device`. All transformations to the tape
before execution should be located in one of these two places. This
wasn't possible previously when adding validation, because the gradient
information wasn't being added to the `ExecutionConfig` passed to
`device.preprocess`, so the transforms couldn't be added in `preprocess`
conditionally based on gradient method.

**Description of the Change:**

- we add a `make_execution_config` function that checks if we are
currently taking a gradient and if so, by what method, and creates an
`ExecutionConfig` that reflects that
- we update `trace_quantum_function` to pass the updated
`ExecutionConfig` to the device's `preprocess` function
- now that the preprocess function has gradient information, we move the
verification steps from `trace_quantum_function` into
`device.preprocess`.

**Benefits:**
The existing gradient-based validation can be moved into `preprocess` to
conform to the PennyLane device API, and future validation and
decomposition transforms based on the gradient can also be added there.
It also resolves a circular import issue so we can remove the
non-top-level imports in `verification.py`.

---------

Co-authored-by: Sergei Mironov <sergei.mironov@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>