[Frontend] Split non-commuting observables (#821)
**Context:**
State-based simulators can generally get the values for non-commuting
observables together on a single execution. Hardware and some simulators
require non-commuting observables to be separated into different
executions.

**Description of the Change:**
We add a `non_commuting_observables` flag to the TOML file. If `True`,
the device can deal with non-commuting observables.

We check the `non_commuting_observables_flag` in `device.preprocess`,
and add the `split_non_commuting` transform to the transform_program if
it's `False`.