Update determine runner workflow to use large runners if the branch name matches the rc branching format (#846)
**Context:**
This PR introduces a change to the CI, where a PR that matches the
naming format of the RC branches (eg: `vX.Y.Z-rcN`) is automatically
on-boarded to use large runner queue *without the usage of a label*

**Description of the Change:**
This change has been introduced within pennylane-lightning with the
following two PRs:
- PennyLaneAI/pennylane-lightning#769
- PennyLaneAI/pennylane-lightning#774

**Benefits:**
Lower wait times for rc branch builds

**Possible Drawbacks:**
As noted in the lightning PRs, though the runners have parity, it is
possible for a workflow to fail on the larger runners. The `urgent`
label will be added to this PR to ensure jobs pass.

**Related GitHub Issues:**