Stop enabling and disabling new_opmath (#845)
**Context:**

In `test_measurements_results.py`, we explicitly enable and then disable
`new_opmath` inside each test. Enabling `new_opmath` is no longer
necessary, because it's the default. Disabling it at the end of each
test means that when the test is completed, we are set to use legacy
`opmath`.

Locally we run each test file separately, but in CI, any tests on the
same runner that are run after this file are run with new_opmath
disabled. This can lead to discrepancies between local and CI testing.

**Description of the Change:**
We remove the explicit enabling and disabling of `new_opmath`

**Benefits:**
We will be testing everything with new opmath in CI as intended, and
local and CI tests will correspond better to each other.

**Possible Drawbacks:**
This came up in a couple of PRs because added functionality failed with
legacy opmath (otherwise this would never have been visible). Without
this, we wouldn't have caught the legacy opmath incompatibility. But if
we want to check for incompatibility with legacy opmath, we should add a
workflow to the test matrix checking the whole thing regularly.