generalize `mitigate_with_zne` to accept arbitrary fitting functions (#806)
**Context:** As laid out in
https://github.com/PennyLaneAI/catalyst/issues/754 we would like to add
exponential fitting capabilities to the ZNE implementation in catalyst.
As a first step, this PR generalizes the `mitigate_with_zne` function to
accept a callable object. Adding exponential fitting after this goes
through review will mean a concise follow on PR to add exponential
fitting capabilities and further tests.

**Description of the Change:** The change can be summarized as
```diff
- def mitigate_with_zne(fn=None, *, scale_factors=None, deg=None):
+ def mitigate_with_zne(fn=None, *, scale_factors=None, extrapolate=None):
```

**Benefits:** This addition makes it easier to add new extrapolation
techniques! The default behavior when (previously) not using the `deg`
argument, and (currently) not using the `extrapolate` argument remain
the same.

**Possible Drawbacks:** The work required by a `mitigate_with_zne` user
is potentially more cumbersome as they can no longer _just_ pass a
polynomial degree, and must specify a function to do the fitting. A
utility is provided, but unsure how we want to expose that to the user
or just point them to
[`pennylane.transforms.poly_extrapolate`](https://docs.pennylane.ai/en/stable/code/api/pennylane.transforms.poly_extrapolate.html)
in the documentation.

**Related GitHub Issues:**
https://github.com/PennyLaneAI/catalyst/issues/754 (this PR is part 1 of
a fix)