Reduce # of configurations for nightly
TestPyPI only provides 10GB of storage for a project, which with our
current configuration would be filled in about ~10 releases. After the
space is filled, older releases need to manually deleted (until better
solution is found).
With a nightly schedule, reducing the number of configurations should
help alleviate the problem somewhat.