Introduce folding argument in ZNE mitigation (#946)
**Context:**
We'd like to extend Zero-noise extrapolation (ZNE) with different
folding techniques. For more details, see #755 and
https://mitiq.readthedocs.io/en/stable/guide/zne-3-options.html#unitary-folding.

**Description of the Change:**
In this PR a `folding` argument is added both to the ZNE mitigation
routine in Catalyst API/frontend/MLIR dialect.
At the moment the argument is passed throughout the stack from the
frontend all the way to the MLIR dialect. However it is mute, in the
sense that it's not being used to change the behavior of the MLIR ZNE
transform.

**Benefits:**
This is an incremental change to discuss design and to accomodate
follow-up pull requests where the argument will be used to branch off
the ZNE mitigation code for different unitary folding techniques.

**Possible Drawbacks:**
The argument is left as `global` by default and it's not yet used in the
MLIR transform, so the current behavior is unchanged. If one tries to
use different value for the new argument, exceptions are raised.

**Related GitHub Issues:**
This is an incremental change towards #755.